requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
sentence-transformers
scikit-learn
numpy
debugpy
kafka-python
requests
python-dotenv
pydantic
redis